refactored PageFetcher class to interface
